CreateInstance: Rename src file
